+++ src/App.js
@@ -1,4 +1,8 @@
 import { useState } from 'react';
+import Logo from './Logo';
+import Form from './Form';
+import PackingList from './PackingList';
+import Stats from './Stats';
 
 const initialItems = [
 	{ id: 1, description: 'Passports', quantity: 2, packed: false },
@@ -45,128 +49,5 @@ export default function App() {
 			/>
 			<Stats items={items} />
 		</div>
-	);
-}
-
-function Logo() {
-	return <h1> 🎄 Far Away 🎁</h1>;
-}
-
-function Form({ onAddItems }) {
-	const [description, setDescription] = useState('');
-	const [number, setNumber] = useState(1);
-
-	function handleSubmit(e) {
-		e.preventDefault();
-
-		if (!description) return;
-
-		const newItem = { description, number, packed: false, id: Date.now() };
-		console.log(newItem);
-
-		onAddItems(newItem);
-
-		setDescription('');
-		setNumber(1);
-	}
-
-	return (
-		<form className='add-form' onSubmit={handleSubmit}>
-			<h3>What Do You Need For Your 😍Trip</h3>
-			<select
-				value={number}
-				onChange={(e) => setNumber(Number(e.target.value))}
-			>
-				{Array.from({ length: 20 }, (_, i) => i + 1).map((num) => (
-					<option value={num} key={num}>
-						{num}
-					</option>
-				))}
-			</select>
-			<input
-				type='text'
-				placeholder='Item....'
-				value={description}
-				onChange={(e) => setDescription(e.target.value)}
-			/>
-			<button>Add</button>
-		</form>
-	);
-}
-
-function PackingList({ items, onDeleteItem, onToggleItem, onClearList }) {
-	const [sortBy, setSortBy] = useState('packed');
-
-	let sortedItems;
-	if (sortBy === 'input') sortedItems = items;
-	if (sortBy === 'description')
-		sortedItems = items
-			.slice()
-			.sort((a, b) => a.description.localeCompare(b.description));
-
-	if (sortBy === 'packed')
-		sortedItems = items
-			.slice()
-			.sort((a, b) => Number(a.packed) - Number(b.packed));
-	return (
-		<div className='list'>
-			<ul>
-				{sortedItems.map((item) => (
-					<Item
-						item={item}
-						onDeleteItem={onDeleteItem}
-						onToggleItem={onToggleItem}
-						key={item.id}
-					/>
-				))}
-			</ul>
-			<div className='actions'>
-				<select value={sortBy} onChange={(e) => setSortBy(e.target.value)}>
-					<option value='input'>Sort By Input order</option>
-					<option value='description'>Sort By Description</option>
-					<option value='packed'>Sort By Packed Status</option>
-				</select>
-				<button onClick={onClearList}> Clear List</button>
-			</div>
-		</div>
-	);
-}
-
-function Item({ item, onDeleteItem, onToggleItem }) {
-	return (
-		<li>
-			<input
-				type='checkbox'
-				value={item.packed}
-				onChange={() => onToggleItem(item.id)}
-			/>{' '}
-			<span style={item.packed ? { textDecoration: 'line-through' } : {}}>
-				{item.quantity} {item.description}
-			</span>
-			<button onClick={() => onDeleteItem(item.id)}>✌</button>
-		</li>
-	);
-}
-
-function Stats({ items }) {
-	if (!items.length)
-		return (
-			<p className='stats'>
-				<em>Start Adding Some Items To Your List</em>
-			</p>
-		);
-	const numItems = items.length;
-	const numPacked = items.filter((item) => item.packed).length;
-	const percentage = Math.round((numPacked / numItems) * 100);
-
-	return (
-		<footer className='stats'>
-			<em>
-				{percentage === 100
-					? 'You Got Have Everythong Ready To Go ✔'
-					: `You Have ${numItems} items on your list and you already packed
-				${numPacked} (${percentage}%)`}
-			</em>
-		</footer>
 	);
 }
