+++ src/PackingList.js
@@ -1,0 +1,45 @@
+import { useState } from 'react';
+import Item from './Item';
+
+export default function PackingList({
+	items,
+	onDeleteItem,
+	onToggleItem,
+	onClearList,
+}) {
+	const [sortBy, setSortBy] = useState('packed');
+
+	let sortedItems;
+	if (sortBy === 'input') sortedItems = items;
+	if (sortBy === 'description')
+		sortedItems = items
+			.slice()
+			.sort((a, b) => a.description.localeCompare(b.description));
+
+	if (sortBy === 'packed')
+		sortedItems = items
+			.slice()
+			.sort((a, b) => Number(a.packed) - Number(b.packed));
+	return (
+		<div className='list'>
+			<ul>
+				{sortedItems.map((item) => (
+					<Item
+						item={item}
+						onDeleteItem={onDeleteItem}
+						onToggleItem={onToggleItem}
+						key={item.id}
+					/>
+				))}
+			</ul>
+			<div className='actions'>
+				<select value={sortBy} onChange={(e) => setSortBy(e.target.value)}>
+					<option value='input'>Sort By Input order</option>
+					<option value='description'>Sort By Description</option>
+					<option value='packed'>Sort By Packed Status</option>
+				</select>
+				<button onClick={onClearList}> Clear List</button>
+			</div>
+		</div>
+	);
+}
